Move all methods from 'main' file to own files
